refactor: extract penalties, reason_codes, and solar modules (#793)
Move penalty/reason/solar methods from DPPlanner to standalone modules:
- engine/penalties.py: get_solar_opportunity_penalty_factor(), get_futile_cycling_penalty_factor()
- engine/reason_codes.py: classify_reason(), classify_hold_reason(), classify_export_reason(),
  classify_charge_reason(), _is_target_shortfall_risk(), _is_cheap_import_window(),
  _is_blind_to_future_solar()
- engine/solar.py: can_solar_reach_target(), projected_solar_soc_gain_pct(),
  projected_solcast_gain_pct(), get_forecast_accuracy()

core.py reduced from ~1280 to ~747 lines.
