+++ tests/engine/test_terminal_cost_accuracy.py
@@ -1,6 +1,7 @@
 from unittest.mock import Mock
 
 from custom_components.localshift.engine.core import DPPlanner
+from custom_components.localshift.engine.solar import get_forecast_accuracy
 
 
 class TestGetForecastAccuracy:
@@ -9,7 +10,7 @@ class TestGetForecastAccuracy:
     def test_no_tracker_returns_one(self):
         """When no tracker exists, return 1.0 (no discount)."""
         planner = DPPlanner.__new__(DPPlanner)
-        result = planner._get_forecast_accuracy(None)
+        result = get_forecast_accuracy(None)
         assert result == 1.0
 
     def test_tracker_returns_100_returns_one(self):
@@ -17,7 +18,7 @@ class TestGetForecastAccuracy:
         tracker = Mock()
         tracker.metrics.accuracy = 100
         planner = DPPlanner.__new__(DPPlanner)
-        result = planner._get_forecast_accuracy(tracker)
+        result = get_forecast_accuracy(tracker)
         assert result == 1.0
 
     def test_tracker_returns_50_returns_point_five(self):
@@ -25,7 +26,7 @@ class TestGetForecastAccuracy:
         tracker = Mock()
         tracker.metrics.accuracy = 50
         planner = DPPlanner.__new__(DPPlanner)
-        result = planner._get_forecast_accuracy(tracker)
+        result = get_forecast_accuracy(tracker)
         assert result == 0.5
 
     def test_tracker_returns_37_returns_point_three_seven(self):
@@ -33,7 +34,7 @@ class TestGetForecastAccuracy:
         tracker = Mock()
         tracker.metrics.accuracy = 37
         planner = DPPlanner.__new__(DPPlanner)
-        result = planner._get_forecast_accuracy(tracker)
+        result = get_forecast_accuracy(tracker)
         assert result == 0.37
 
     def test_tracker_returns_none_returns_one(self):
@@ -41,7 +42,7 @@ class TestGetForecastAccuracy:
         tracker = Mock()
         tracker.metrics.accuracy = None
         planner = DPPlanner.__new__(DPPlanner)
-        result = planner._get_forecast_accuracy(tracker)
+        result = get_forecast_accuracy(tracker)
         assert result == 1.0
 
     def test_tracker_returns_zero_returns_one(self):
@@ -49,7 +50,7 @@ class TestGetForecastAccuracy:
         tracker = Mock()
         tracker.metrics.accuracy = 0
         planner = DPPlanner.__new__(DPPlanner)
-        result = planner._get_forecast_accuracy(tracker)
+        result = get_forecast_accuracy(tracker)
         assert result == 1.0
 
     def test_tracker_returns_negative_returns_one(self):
@@ -57,14 +58,14 @@ class TestGetForecastAccuracy:
         tracker = Mock()
         tracker.metrics.accuracy = -10
         planner = DPPlanner.__new__(DPPlanner)
-        result = planner._get_forecast_accuracy(tracker)
+        result = get_forecast_accuracy(tracker)
         assert result == 1.0
 
     def test_tracker_no_metrics_attribute_returns_one(self):
         """When tracker has no metrics attribute, return 1.0."""
         tracker = Mock(spec=[])  # No attributes at all
         planner = DPPlanner.__new__(DPPlanner)
-        result = planner._get_forecast_accuracy(tracker)
+        result = get_forecast_accuracy(tracker)
         assert result == 1.0
